Move libpqwalproposer to walproposer_pg
