+++ pgxn/neon/Makefile
@@ -7,7 +7,6 @@ OBJS = \
 	extension_server.o \
 	file_cache.o \
 	libpagestore.o \
-	libpqwalproposer.o \
 	neon.o \
 	pagestore_smgr.o \
 	relsize_cache.o \
